implemented native encoded value width for all integers

+++ os/cfe.c
@@ -96,7 +96,7 @@ int bplib_os_log(const char* file, unsigned int line, int error, const char* fmt
 /*--------------------------------------------------------------------------------------
  * bplib_os_systime - returns seconds
  *-------------------------------------------------------------------------------------*/
-uint32_t bplib_os_systime(void)
+unsigned long bplib_os_systime(void)
 {
     CFE_TIME_SysTime_t sys_time = CFE_TIME_GetTime();
     return sys_time.Seconds - BPLIB_CFE_SECS_AT_2000;
